refactor into burn-to-mint package
