+++ Cargo.toml
@@ -56,6 +56,7 @@ open-edition-minter  = { version = "3.1.0", path = "contracts/minters/open-editi
 whitelist-immutable  = { version = "3.1.0", path = "contracts/whitelists/whitelist-immutable" }
 sg-whitelist-flex    = { version = "3.1.0", path = "contracts/whitelists/whitelist-flex" }
 ethereum-verify      = { version = "3.1.0", path = "packages/ethereum-verify" }
+burn-to-mint         = { version = "3.1.0", path = "packages/burn-to-mint" }
 sg-eth-airdrop       = { version = "3.1.0", path = "contracts/sg-eth-airdrop" }
 test-suite           = { version = "3.1.0", path = "test-suite" }
 semver               = "1"
