+++ contracts/minters/open-edition-minter/Cargo.toml
@@ -43,4 +43,5 @@ sg-metadata             = { workspace = true }
 thiserror               = { workspace = true }
 url                     = { workspace = true }
 open-edition-factory    = { workspace = true, features = ["library"] }
-semver                  = {workspace = true }
+semver                  = { workspace = true }
+burn-to-mint            = { workspace = true }
