+++ contracts/minters/open-edition-minter/src/contract.rs
@@ -1,8 +1,8 @@
 #[cfg(not(feature = "library"))]
 use cosmwasm_std::entry_point;
 use cosmwasm_std::{
-    coin, to_binary, Addr, BankMsg, Binary, Coin, CosmosMsg, Decimal, Deps, DepsMut, Empty, Env,
-    MessageInfo, Order, Reply, ReplyOn, StdError, StdResult, Timestamp, Uint128, WasmMsg,
+    coin, to_binary, Addr, BankMsg, Binary, Coin, Decimal, Deps, DepsMut, Empty, Env, MessageInfo,
+    Order, Reply, ReplyOn, StdError, StdResult, Timestamp, Uint128, WasmMsg,
 };
 use cw2::set_contract_version;
 use cw721::Cw721ReceiveMsg;
@@ -187,7 +187,7 @@ pub fn execute(
             execute_update_per_address_limit(deps, env, info, per_address_limit)
         }
         ExecuteMsg::MintTo { recipient } => execute_mint_to(deps, env, info, recipient),
-        ExecuteMsg::ReceiveNft(msg) => execute_burn_to_mint(deps, env, info, msg),
+        ExecuteMsg::ReceiveNft(msg) => call_burn_to_mint(deps, env, info, msg),
     }
 }
 
@@ -680,33 +680,16 @@ pub fn execute_update_per_address_limit(
         .add_attribute("limit", per_address_limit.to_string()))
 }
 
-pub fn execute_burn_to_mint(
+pub fn call_burn_to_mint(
     _deps: DepsMut,
     env: Env,
     info: MessageInfo,
     msg: Cw721ReceiveMsg,
 ) -> Result<Response, ContractError> {
-    let mut res = Response::new();
-    let burn_msg = cw721::Cw721ExecuteMsg::Burn {
-        token_id: msg.token_id,
-    };
-    let cosmos_burn_msg = CosmosMsg::Wasm(WasmMsg::Execute {
-        contract_addr: info.sender.to_string(),
-        msg: to_binary(&burn_msg)?,
-        funds: vec![],
-    });
-    res = res.add_message(cosmos_burn_msg);
-
     let execute_mint_msg = ExecuteMsg::Mint {};
-    let cosmos_mint_msg = CosmosMsg::Wasm(WasmMsg::Execute {
-        contract_addr: env.contract.address.to_string(),
-        msg: to_binary(&execute_mint_msg)?,
-        funds: vec![],
-    });
-    let res = res.add_message(cosmos_mint_msg);
-    Ok(res)
+    burn_to_mint::generate_burn_mint_response(info, env, msg, execute_mint_msg)
+        .map_err(|_| ContractError::BurnToMintError {})
 }
-
 // if admin_no_fee => no fee,
 // else if in whitelist => whitelist price
 // else => config unit price
