+++ contracts/minters/vending-minter/Cargo.toml
@@ -51,4 +51,5 @@ sg-whitelist    = { workspace = true, features = ["library"] }
 thiserror       = { workspace = true }
 url             = { workspace = true }
 vending-factory = { workspace = true, features = ["library"] }
-semver           = {workspace = true  }
+semver          = {workspace = true  }
+burn-to-mint    = { workspace = true }
