+++ contracts/minters/vending-minter/src/error.rs
@@ -107,4 +107,7 @@ pub enum ContractError {
 
     #[error("Expected fungible token, received nonfungible")]
     IncorrectFungibility {},
+
+    #[error("Burn to Mint Error")]
+    BurnToMintError {},
 }
