+++ packages/burn-to-mint/src/msg.rs
@@ -1,0 +1,6 @@
+use cosmwasm_schema::cw_serde;
+
+#[cw_serde]
+pub struct TokenUriMsg {
+    pub token_uri: String,
+}
